Rearranging the order of energy sources so that intermittent is on top
This is the proposed edit for issue https://github.com/tmrowco/electricitymap-contrib/issues/2522 to make it easier to see how intermittency can mess up the ramping up capabilities of dispatchable energy sources (in the middle). This is called "the duck curve". Wind and solar is at the top, and hydropower is in the same category as dispatchable sources (gas and oil).